Refactored modal into seperate component
